Existing tests work after restructuring
